fix(e2e): make free-agency-submission the sole ibl_fa_offers owner
Master CI shard-2 was failing intermittently in free-agency-submission
("amend"/"submit max contract": Delete button absent on readback). Root
cause: under Playwright fullyParallel sharding, multiple spec files mutate
the GLOBAL ibl_fa_offers table in different workers concurrently. A
table-wide wipe (clear_offers / clear-fa-offers, count→0) races against the
submission spec's in-flight pid=11 offer and deletes it mid-test. PR #884
added spec files, which reshuffled deterministic shard boundaries and
co-located free-agency-admin-clear-offers.spec.ts with the submission spec
in shard 2 — exposing a latent fragility, not a code regression.

Fix: consolidate ALL global ibl_fa_offers mutation into one serial file
(free-agency-submission.spec.ts), which runs on a single worker. No other
file can then touch the table concurrently regardless of how tests shard.

- Merge free-agency-admin-clear-offers.spec.ts (clear_offers) into the
  submission spec; delete the old file.
- Move free-agency.spec.ts's "validation errors" block (submits pid=11
  offers + clears the table) into the submission spec. free-agency.spec.ts
  is now read-only w.r.t. offers.
- Drop csrf-rejection.spec.ts's reset-fa-offers hooks and the redundant
  global offer-count readback. The forged token is rejected before the
  clear_offers handler runs, so "Security validation failed" is itself
  proof of no mutation — the spec no longer needs to touch the table.
- Add bin/check-e2e-fa-offers-owner (wired into the E2E Hygiene Gate) to
  fail CI if any non-owner spec reintroduces global FA-offer mutation.

Verified: deterministic per-shard check confirms the submission spec is the
sole ibl_fa_offers mutator in every shard; 3 consecutive green runs of all
3 FA specs at --workers=4 against the worktree Docker stack.
